+++ TablePro/Models/Database/TableFilter.swift
@@ -187,12 +187,30 @@ struct TabFilterState: Equatable, Hashable, Codable {
     var commit: FilterCommit?
     var isVisible: Bool
     var filterLogicMode: FilterLogicMode
+    var keyPattern: String
+    var keyTypeScope: String?
 
     init(isVisible: Bool = false) {
         self.filters = []
         self.commit = nil
         self.isVisible = isVisible
         self.filterLogicMode = .and
+        self.keyPattern = ""
+        self.keyTypeScope = nil
+    }
+
+    enum CodingKeys: String, CodingKey {
+        case filters, commit, isVisible, filterLogicMode, keyPattern, keyTypeScope
+    }
+
+    init(from decoder: Decoder) throws {
+        let container = try decoder.container(keyedBy: CodingKeys.self)
+        self.filters = try container.decodeIfPresent([TableFilter].self, forKey: .filters) ?? []
+        self.commit = try container.decodeIfPresent(FilterCommit.self, forKey: .commit)
+        self.isVisible = try container.decodeIfPresent(Bool.self, forKey: .isVisible) ?? false
+        self.filterLogicMode = try container.decodeIfPresent(FilterLogicMode.self, forKey: .filterLogicMode) ?? .and
+        self.keyPattern = try container.decodeIfPresent(String.self, forKey: .keyPattern) ?? ""
+        self.keyTypeScope = try container.decodeIfPresent(String.self, forKey: .keyTypeScope)
     }
 
     var appliedFilters: [TableFilter] {